build: don't delete toolchain during clean, add distclean instead
The toolchain is large and takes a while to extract (xz is slow), and
there's no need to delete and re-extract it between builds, unlike the
source code.

Add a "distclean" action which does delete the toolchain, as well as the
toolchain and kernel source tarballs, and make KERNEL_TAR a global
variable set up in init_vars so it can be re-used.